Split ev/providers/tools.py into a package by domain
The 956-line monolith mixed weather, web search, maps, calendar and
email in one file with two duplicate/dead definitions (rain_tomorrow,
fetch_text — each shadowed by a later duplicate). Split into
ev/providers/tools/{weather,websearch,google_auth,maps,calendar,email}.py,
re-exported from __init__.py so `from ev.providers import tools` and
every tools.<name>(...) call site keeps working unchanged. Dropped the
dead duplicate definitions and fixed a latent NameError in
nearby_places (missing `import re`) that would only surface on a
free-text place search.
