+++ ev/providers/tools/calendar.py
@@ -1,0 +1,96 @@
+"""Google Calendar tools."""
+
+from __future__ import annotations
+
+import logging
+
+from .google_auth import _google_service
+
+log = logging.getLogger("ev.tools")
+
+
+def calendar_upcoming(config, account: str, max_results: int = 5) -> str:
+    """List the user's upcoming Google Calendar events."""
+    from datetime import datetime, timezone
+
+    try:
+        service = _google_service(config, account, "calendar", "v3")
+        now = datetime.now(timezone.utc).isoformat()
+        events = (
+            service.events()
+            .list(
+                calendarId="primary",
+                timeMin=now,
+                maxResults=max_results,
+                singleEvents=True,
+                orderBy="startTime",
+            )
+            .execute()
+            .get("items", [])
+        )
+    except Exception as exc:
+        log.warning("calendar_upcoming failed (%s)", exc)
+        return f"não consegui acessar a agenda ({exc})"
+
+    if not events:
+        return "nenhum evento próximo na agenda."
+    lines = []
+    for e in events:
+        start = e.get("start", {}).get("dateTime") or e.get("start", {}).get("date")
+        lines.append(f"- {start}: {e.get('summary', '(sem título)')}")
+    return "\n".join(lines)
+
+
+def calendar_list_range(
+    config, account: str, start_iso: str, end_iso: str, max_results: int = 250
+) -> list[dict]:
+    """List Google Calendar events between start and end as structured dicts."""
+    service = _google_service(config, account, "calendar", "v3")
+    events = (
+        service.events()
+        .list(
+            calendarId="primary", timeMin=start_iso, timeMax=end_iso,
+            maxResults=max_results, singleEvents=True, orderBy="startTime",
+        )
+        .execute()
+        .get("items", [])
+    )
+    out = []
+    for e in events:
+        s, en = e.get("start", {}), e.get("end", {})
+        out.append({
+            "id": e.get("id"),
+            "summary": e.get("summary", "(sem título)"),
+            "start": s.get("dateTime") or s.get("date"),
+            "end": en.get("dateTime") or en.get("date"),
+            "all_day": "date" in s and "dateTime" not in s,
+            "link": e.get("htmlLink"),
+        })
+    return out
+
+
+def calendar_delete(config, account: str, event_id: str) -> bool:
+    """Delete a Google Calendar event by id."""
+    service = _google_service(config, account, "calendar", "v3")
+    service.events().delete(calendarId="primary", eventId=event_id).execute()
+    return True
+
+
+def calendar_create(
+    config, account: str, summary: str, start_iso: str, end_iso: str
+) -> str:
+    """Create a Google Calendar event."""
+    try:
+        service = _google_service(config, account, "calendar", "v3")
+        event = {
+            "summary": summary,
+            "start": {"dateTime": start_iso},
+            "end": {"dateTime": end_iso},
+        }
+        created = (
+            service.events().insert(calendarId="primary", body=event).execute()
+        )
+        return f"evento criado: {created.get('htmlLink', summary)}"
+    except Exception as exc:
+        log.warning("calendar_create failed (%s)", exc)
+        return f"não consegui criar o evento ({exc})"
